requirements.txt
beautifulsoup4==4.6.3
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
html5lib==1.0.1
idna==2.8
lxml==4.2.5
Pillow==5.3.0
progress==1.4
pycparser==2.19
Pyphen==0.9.5
PyYAML==3.13
requests==2.21.0
six==1.12.0
tinycss2==0.6.1
urllib3==1.24.1
WeasyPrint==43
webencodings==0.5.1
